--- a/Dockerfile
+++ b/Dockerfile
@@ -73,9 +73,9 @@ ENTRYPOINT ["python3", "-m", "vllm.entrypoints.api_server"]
 
 # openai api server alternative
 FROM vllm-base AS vllm-openai
-# install additional dependencies for openai api server, and mixtral
+# install additional dependencies for openai api server
 RUN --mount=type=cache,target=/root/.cache/pip \
-    pip install accelerate megablocks
+    pip install accelerate fschat
 
 COPY --from=build /workspace/vllm/*.so /workspace/vllm/
 COPY vllm vllm