This PR implements a more efficient parallelism scheme for the Mixtral model. Instead of sharding the layers of each expert by rank, we instead shard whole experts across ranks. This gives us several benefits:

- We reduce the amount of communication between ranks
- We do not require megablocks (meaning we can now support non-CUDA accelerators)
- The operations become more efficient and CUDA-graphable.

In the new design, each expert will conduct a dense matrix multiplication of the whole batch, and then rows not assigned to the expert will be zeroed out before accumulation. This results in a slight inefficiency for tensor parallel sizes below the number of experts - it means that we will essentially always do the upper performance bound computation. However, we have not found this to be an issue in practice. A potential improvement would be to use a sparse/grouped GEMM kernel (at least for prefill - for decode it shouldn't matter).

We have benchmarked this change and found that it lowers the e2e latency for Mixtral by 4x-5x on A100-40GB TP8 compared to the previous implementation.

Furthermore, the PR refactors the Mixtral model for compatibility with Hugging Face format and safetensor weights, and adds quantization support.