@WoosukKwon 2023-12-13
Hi @Yard1, thanks for the amazing work!

I've just tested the PR on `examples/llm_engine_example.py` and got the following results:
* Current main
```
INFO 12-13 23:31:55 llm_engine.py:222] # GPU blocks: 86172, # CPU blocks: 8192
INFO 12-13 23:31:58 llm_engine.py:649] Avg prompt throughput: 0.0 tokens/s, Avg generation throughput: 0.0 tokens/s, Running: 1 reqs, Swapped: 0 reqs, Pending: 0 reqs, GPU KV cache usage: 0.0%, CPU KV cache usage: 0.0%
RequestOutput(request_id=0, prompt='A robot may not injure a human being', prompt_token_ids=[1, 330, 18401, 993, 459, 5891, 482, 264, 2930, 1250], prompt_logprobs=[None, {330: -9.912246704101562, 22478: -0.7872462272644043}, {18401: -8.597543716430664, 633: -3.347543478012085}, {993: -4.565238952636719, 369: -2.1902387142181396}, {459: -0.4373227059841156}, {5891: -0.4258776903152466}, {482: -3.099436753473128e-06}, {264: -0.0011317284079268575}, {2930: -0.0006484074983745813}, {1250: -0.009901456534862518}], outputs=[CompletionOutput(index=0, text=' or, through inaction, allow a human being to come to harm.\n', token_ids=[442, 28725, 1059, 297, 1774, 28725, 1914, 264, 2930, 1250, 298, 1567, 298, 6241, 28723, 13], cumulative_logprob=-0.6244106972517329, logprobs=[{442: -0.017248855903744698}, {28725: -0.002303091809153557}, {1059: -0.0011830481234937906}, {297: -0.00041952868923544884}, {1774: -7.164221460698172e-05}, {28725: -0.0003152588615193963}, {1914: -0.0006347072194330394}, {264: -0.0005576247931458056}, {2930: -0.00010775939153973013}, {1250: -0.0015303102554753423}, {298: -0.0005830018781125546}, {1567: -0.0004058252670802176}, {298: -0.0002112165529979393}, {6241: -0.0003516055876389146}, {28723: -0.03660520166158676}, {13: -0.5618820190429688}], finish_reason=length)], finished=True)
RequestOutput(request_id=1, prompt='To be or not to be,', prompt_token_ids=[1, 1791, 347, 442, 459, 298, 347, 28725], prompt_logprobs=None, outputs=[CompletionOutput(index=0, text=' that is the question.\nWhether ’tis nobler in the mind', token_ids=[369, 349, 272, 2996, 28723, 13, 23842, 620, 24978, 28707, 278, 7169, 1523, 297, 272, 2273], cumulative_logprob=-5.713744854774632, logprobs=None, finish_reason=length)], finished=True)
RequestOutput(request_id=2, prompt='What is the meaning of life?', prompt_token_ids=[1, 1824, 349, 272, 5746, 302, 1411, 28804], prompt_logprobs=None, outputs=[CompletionOutput(index=0, text='\n\nThe meaning of life is the question of the purpose and significance of life', token_ids=[13, 13, 1014, 5746, 302, 1411, 349, 272, 2996, 302, 272, 6032, 304, 18309, 302, 1411], cumulative_logprob=-8.794605396687984, logprobs=None, finish_reason=length), CompletionOutput(index=3, text=' It’s a question that’s been asked by philosophers, theolog', token_ids=[661, 28809, 28713, 264, 2996, 369, 28809, 28713, 750, 2261, 486, 8829, 404, 28725, 272, 1165], cumulative_logprob=-9.33446236141026, logprobs=None, finish_reason=length)], finished=True)
RequestOutput(request_id=3, prompt='It is only with the heart that one can see rightly', prompt_token_ids=[1, 661, 349, 865, 395, 272, 3031, 369, 624, 541, 1032, 1103, 346], prompt_logprobs=None, outputs=[CompletionOutput(index=0, text='; what is essential is invisible to the eye.\n\nAntoine de Saint', token_ids=[28745, 767, 349, 7974, 349, 20187, 298, 272, 5421, 28723, 13, 13, 13389, 21265, 340, 6393], cumulative_logprob=-2.537341303512221, logprobs=None, finish_reason=length), CompletionOutput(index=1, text='; what is essential is invisible to the eye. Antoine de Saint-Ex', token_ids=[28745, 767, 349, 7974, 349, 20187, 298, 272, 5421, 28723, 3821, 21265, 340, 6393, 28733, 966], cumulative_logprob=-2.979412608925486, logprobs=None, finish_reason=length), CompletionOutput(index=2, text='; what is essential is invisible to the eye. – Antoine de Saint-', token_ids=[28745, 767, 349, 7974, 349, 20187, 298, 272, 5421, 28723, 764, 3821, 21265, 340, 6393, 28733], cumulative_logprob=-3.1470024501613807, logprobs=None, finish_reason=length)], finished=True)
```

* This PR
```
INFO 12-13 23:20:14 llm_engine.py:222] # GPU blocks: 57756, # CPU blocks: 819295MB/s]
INFO 12-13 23:20:17 llm_engine.py:649] Avg prompt throughput: 0.0 tokens/s, Avg generation throughput: 0.0 tokens/s, Running: 1 reqs, Swapped: 0 reqs, Pending: 0 reqs, GPU KV cache usage: 0.0%, CPU KV cache usage: 0.0%
RequestOutput(request_id=0, prompt='A robot may not injure a human being', prompt_token_ids=[1, 330, 18401, 993, 459, 5891, 482, 264, 2930, 1250], prompt_logprobs=[None, {330: -9.617840766906738, 12: -1.7154966592788696}, {18401: -8.787067413330078, 330: -2.7558176517486572}, {993: -4.204432010650635, 349: -2.0169320106506348}, {459: -0.3415136933326721}, {5891: -1.0073399543762207, 6241: -0.5073400139808655}, {482: -1.3708974620385561e-05}, {264: -0.1135331317782402}, {2930: -0.002309514442458749}, {1250: -0.016736455261707306}], outputs=[CompletionOutput(index=0, text=', or, more importantly, a robot may not kill a human being.\n', token_ids=[28725, 442, 28725, 680, 21485, 28725, 264, 18401, 993, 459, 4015, 264, 2930, 1250, 28723, 13], cumulative_logprob=-7.275053498335183, logprobs=[{28725: -0.16343587636947632}, {442: -0.21259483695030212}, {28725: -0.1041431725025177}, {680: -1.0776935815811157}, {21485: -0.2229764610528946}, {28725: -0.01339601818472147}, {264: -1.1102567911148071}, {18401: -0.1942392736673355}, {993: -0.3014945983886719}, {459: -0.05710757523775101}, {4015: -1.3823846578598022}, {264: -0.5338531732559204}, {2930: -0.08587013930082321}, {1250: -0.040455106645822525}, {28723: -0.33675137162208557}, {13: -1.4384008646011353}], finish_reason=length)], finished=True)
RequestOutput(request_id=1, prompt='To be or not to be,', prompt_token_ids=[1, 1791, 347, 442, 459, 298, 347, 28725], prompt_logprobs=None, outputs=[CompletionOutput(index=0, text=' that is the question of every person’s life.\nTo live, to', token_ids=[369, 349, 272, 2996, 302, 1012, 1338, 28809, 28713, 1411, 28723, 13, 1551, 2943, 28725, 298], cumulative_logprob=-19.226570382204045, logprobs=None, finish_reason=length)], finished=True)
RequestOutput(request_id=2, prompt='What is the meaning of life?', prompt_token_ids=[1, 1824, 349, 272, 5746, 302, 1411, 28804], prompt_logprobs=None, outputs=[CompletionOutput(index=0, text='\n\nThe meaning of life is the meaning of one’s life. That', token_ids=[13, 13, 1014, 5746, 302, 1411, 349, 272, 5746, 302, 624, 28809, 28713, 1411, 28723, 1725], cumulative_logprob=-16.94498591311276, logprobs=None, finish_reason=length), CompletionOutput(index=4, text='\n\nThis question was often asked in the ancient and modern days.\n\n', token_ids=[13, 13, 3260, 2996, 403, 2608, 2261, 297, 272, 9467, 304, 4638, 2202, 28723, 13, 13], cumulative_logprob=-28.903032392263412, logprobs=None, finish_reason=length)], finished=True)
RequestOutput(request_id=3, prompt='It is only with the heart that one can see rightly', prompt_token_ids=[1, 661, 349, 865, 395, 272, 3031, 369, 624, 541, 1032, 1103, 346], prompt_logprobs=None, outputs=[CompletionOutput(index=1, text='; the\nreasonable world does not know in unces.\n\nHow', token_ids=[28745, 272, 13, 14991, 522, 1526, 1235, 459, 873, 297, 521, 1377, 28723, 13, 13, 5660], cumulative_logprob=-11.229475471191108, logprobs=None, finish_reason=length), CompletionOutput(index=0, text='; the\nreasonable world does not know in unces, what the\n', token_ids=[28745, 272, 13, 14991, 522, 1526, 1235, 459, 873, 297, 521, 1377, 28725, 767, 272, 13], cumulative_logprob=-11.59051242750138, logprobs=None, finish_reason=length), CompletionOutput(index=2, text='; the\nreasonable world does not know in unces.\n\nFor', token_ids=[28745, 272, 13, 14991, 522, 1526, 1235, 459, 873, 297, 521, 1377, 28723, 13, 13, 2565], cumulative_logprob=-11.729475471191108, logprobs=None, finish_reason=length)], finished=True)
```

In summary, 1) the results do not match; I feel the current main's output looks more correct, and 2) There's a huge decrease in allocated the KV cache size. Does this mean that this implementation has very high memory overhead?